refactor: extract base vaadin-select logic to reusable mixin
