+++ packages/select/package.json
@@ -37,6 +37,7 @@
     "polymer"
   ],
   "dependencies": {
+    "@open-wc/dedupe-mixin": "^1.3.0",
     "@polymer/polymer": "^3.2.0",
     "@vaadin/a11y-base": "24.2.0-alpha4",
     "@vaadin/button": "24.2.0-alpha4",
